+++ changelog.md
@@ -9,6 +9,7 @@ All notable changes to `TestAssist` will be documented in this file.
 - Assertions for whether a model is inserted, updated or deleted in the database.
 - Moved `FabricateModels` to the same `Database` namespace.
 - Added `assertIsCalled` to assert whether a class is called during a given callback function.
+- Added `disableExceptionHandling` and `withExceptionHandling`.
 
 ## Version 1.0